Separate xarray utility functions from rest
