+++ weldx/util/__init__.py
@@ -1,0 +1,2 @@
+from .util import *
+from .xarray import *
